Group things by tags, use hosts to select tags.
This prepares these dotfiles for use in contexts where I don't want to
install everything, e.g. in a GitHub Codespace I want my development setup,
but not my email setup.

Previously we had a lot of defaults, and then host-specific config for a few
different machines. This had already got pretty messy over time.

The new approach in this commit is to only use host-specific config for rcrc
files, so that each host can be associated with some specific tags.